serverless/requirements.txt
runpod>=1.6.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
